chartedaismonitor/backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
shapely
httpx